* commit '1d7bce46912096efada0443182295ffe0f2cf077':
  居中布局性能优化
  bug
  bug
  bug
  bug
  bug
  bug
  BI BUG
  指标弹出明细表
  bug
  过滤取字段
  group添加position属性test
  树控件多对多逻辑
  bug
  下拉列表灰化图标处理 数控件多对多取数
  删除无用代码

+++ fbi_base/src/com/fr/bi/stable/utils/BIDBUtils.java
@@ -519,12 +519,12 @@ public class BIDBUtils {
     }
 
 
-    private static long dealWithResultSet(ResultSet rs,
+    private static int dealWithResultSet(ResultSet rs,
                                           DBField[] columns,
                                           Traversal<BIDataValue> traversal,
                                           boolean needCharSetConvert,
                                           String originalCharSetName,
-                                          String newCharSetName, long row) throws SQLException {
+                                          String newCharSetName, int row) throws SQLException {
         @SuppressWarnings("rawtypes")
         DBDealer[] dealers = createDBDealer(needCharSetConvert, originalCharSetName, newCharSetName, columns);
         int ilen = dealers.length;
@@ -539,7 +539,7 @@ public class BIDBUtils {
     }
 
     public static long runSQL(SQLStatement sql, DBField[] columns, Traversal<BIDataValue> traversal) {
-        return runSQL(sql, columns, traversal, 0L);
+        return runSQL(sql, columns, traversal, 0);
     }
 
     /**
@@ -547,7 +547,7 @@ public class BIDBUtils {
      *
      * @param traversal
      */
-    public static long runSQL(SQLStatement sql, DBField[] columns, Traversal<BIDataValue> traversal, long row) {
+    public static int runSQL(SQLStatement sql, DBField[] columns, Traversal<BIDataValue> traversal, int row) {
         com.fr.data.impl.Connection connection = sql.getConn();
         Connection conn = null;
         Statement stmt = null;
